Version gate the storage engine
