+++ server/src/storage/v1/macros.rs
