+++ server/src/storage/v1/unflush.rs
@@ -33,11 +33,11 @@ use crate::corestore::memstore::Keyspace;
 use crate::corestore::memstore::Memstore;
 use crate::corestore::memstore::ObjectID;
 use crate::corestore::table::Table;
-use crate::storage::de::DeserializeInto;
-use crate::storage::flush::Autoflush;
-use crate::storage::interface::DIR_KSROOT;
-use crate::storage::preload::LoadedPartfile;
-use crate::storage::Coremap;
+use crate::storage::v1::de::DeserializeInto;
+use crate::storage::v1::flush::Autoflush;
+use crate::storage::v1::interface::DIR_KSROOT;
+use crate::storage::v1::preload::LoadedPartfile;
+use crate::storage::v1::Coremap;
 use crate::IoResult;
 use core::mem::transmute;
 use std::fs;
